revert: remove empty interface stubs, keep autosave rename
Removes the empty interface stubs from #235 that inflated api:compare
without adding real value. Keeps the autosave.ts → autosave-association.ts
rename since that's a real file placement fix.
